fix: environment not passed to icerberg jar breaking IRSA
IRSA relies on the environment variables AWS_ROLE_ARN and
AWS_WEB_IDENTITY_TOKEN_FILE. We should be passing these to the child
process to allow working in Kubernetes environments without using
hardcoded credentials. This also enables other vars such as
JAVA_TOOL_OPTS to be used as needed. I set an allowlist of variable
prefixes to maintain the security feature of not passing the whole
environment to icerberg.